[FIX/#139] 유저 인증 여부 API 위치 변경
