+++ data/src/main/java/kr/genti/data/dataSource/CreateDataSource.kt
@@ -11,6 +11,4 @@ interface CreateDataSource {
     suspend fun getMultiS3Url(request: List<S3RequestDto>): BaseResponse<List<S3PresignedUrlDto>>
 
     suspend fun postToCreate(request: CreateRequestDto): BaseResponse<Boolean>
-
-    suspend fun getIsUserVerified(): BaseResponse<Boolean>
 }
