+++ data/src/main/java/kr/genti/data/dataSourceImpl/CreateDataSourceImpl.kt
@@ -19,6 +19,4 @@ data class CreateDataSourceImpl
             createService.getMultiS3Url(request)
 
         override suspend fun postToCreate(request: CreateRequestDto): BaseResponse<Boolean> = createService.postToCreate(request)
-
-        override suspend fun getIsUserVerified(): BaseResponse<Boolean> = createService.getIsUserVerified()
     }
