+++ data/src/main/java/kr/genti/data/repositoryImpl/CreateRepositoryImpl.kt
@@ -28,9 +28,4 @@ class CreateRepositoryImpl
             runCatching {
                 createDataSource.postToCreate(request.toDto()).response
             }
-
-        override suspend fun getIsUserVerified(): Result<Boolean> =
-            runCatching {
-                createDataSource.getIsUserVerified().response
-            }
     }
