+++ data/src/main/java/kr/genti/data/service/GenerateService.kt
@@ -46,4 +46,7 @@ interface GenerateService {
 
     @GET("api/v1/open-chat")
     suspend fun getOpenchatData(): BaseResponse<OpenchatDto>
+
+    @GET("api/v1/user-verification")
+    suspend fun getIsUserVerified(): BaseResponse<Boolean>
 }
